# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Gin swagger 示例项目",
        "title": "Gin swagger",
        "contact": {
            "name": "youngxhu",
            "url": "https://youngxhui.top",
            "email": "youngxhui@g mail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
        "/article": {
            "post": {
                "consumes": [
                    "application/x-json-stream"
                ],
                "tags": [
                    "文章"
                ],
                "summary": "提交新的文章内容",
                "operationId": "1",
                "parameters": [
                    {
                        "description": "文章",
                        "name": "article",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ArticleModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ResultModel"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/model.ResultModel"
                        }
                    }
                }
            }
        },
        "/article/{id}": {
            "get": {
                "consumes": [
                    "application/x-json-stream"
                ],
                "summary": "通过文章 id 获取单个文章内容",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ResultModel"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/x-json-stream"
                ],
                "summary": "通过id删除指定文章类型",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ResultModel"
                        }
                    }
                }
            }
        },
        "/articles": {
            "get": {
                "consumes": [
                    "application/x-json-stream"
                ],
                "summary": "获取所有的文章",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ResultModel"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.ArticleModel": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "model.ResultModel": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string",
                    "example": "请求信息"
                }
            }
        }
    }
}